+++ src/hardware/wifi.h
@@ -44,7 +44,13 @@ class Security {
     Q_GADGET
 
  public:
-    enum Enum { OPEN, WPA_PSK, WPA_EAP, WPA2_PSK, WPA2_EAP };
+    /**
+     * WiFi security types.
+     *
+     * OPEN, WPA_PSK, WPA2_WPA3 and WPA3_SAE can be requested when adding a network, AUTO leaves the
+     * choice to the core. The remaining values are only used to classify an existing connection.
+     */
+    enum Enum { OPEN, WPA_PSK, WPA_EAP, WPA2_PSK, WPA2_EAP, WPA2_WPA3, WPA3_SAE, AUTO };
     Q_ENUM(Enum)
 
  private:
@@ -120,8 +126,8 @@ class Wifi : public QObject {
 
     Q_INVOKABLE void turnOn();
     Q_INVOKABLE void turnOff();
-    Q_INVOKABLE void connect(const QString &ssid, const QString &password, uc::hw::Security::Enum security,
-                             bool hidden = false);
+    Q_INVOKABLE void connect(const QString &ssid, const QString &password,
+                             uc::hw::Security::Enum security = uc::hw::Security::AUTO, bool hidden = false);
     Q_INVOKABLE void connectSavedNetwork(int id);
     Q_INVOKABLE void enableSavedNetwork(int id, bool enable);
     Q_INVOKABLE void disconnect();
@@ -139,8 +145,10 @@ class Wifi : public QObject {
     Q_INVOKABLE QString getLastConnectedSsid() { return m_lastConnectedSSid; }
     Q_INVOKABLE QString getLastConnectedPassword() { return m_lastConnectedPassword; }
 
-    void addNetwork(const QString &ssid, const QString &password, uc::hw::Security::Enum security,
-                    bool hidden = false);
+    void addNetwork(const QString &ssid, const QString &password,
+                    uc::hw::Security::Enum security = uc::hw::Security::AUTO, bool hidden = false);
+
+    static core::WifiEnums::WifiSecurity toApiSecurity(uc::hw::Security::Enum security);
     void wifiNetworkCommand(int networkId, core::WifiEnums::WifiNetworkCmd command);
     void wifiCommand(core::WifiEnums::WifiCmd command);
     void clearKnownNetworkList();